chore(components): nav bar clean up, adds links
